requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2
alembic
psycopg2-binary
google-auth
google-api-python-client
python-dotenv
pydantic
pydantic-settings
pyngrok